Make destination a u8*

+++ trunk/linux-2.6.x/arch/armnommu/mach-nds/arm7/spi.c
@@ -53,7 +53,7 @@ void poweroff( void )
 
 }
 
-void read_firmware( u32 address, u16 * destination, int count )
+void read_firmware( u32 address, u8 * destination, int count )
 {
     int i;
